+++ xcube_cci/cciodp.py
@@ -37,7 +37,7 @@ import urllib.parse
 import warnings
 from datetime import datetime
 from dateutil.relativedelta import relativedelta
-from typing import List, Dict, Tuple, Optional, Union
+from typing import List, Dict, Tuple, Optional, Union, Mapping
 from urllib.parse import quote
 from xcube_cci.constants import CCI_ODD_URL
 from xcube_cci.constants import DEFAULT_NUM_RETRIES
@@ -604,42 +604,45 @@ class CciOdp:
                start_date: Optional[str] = None,
                end_date: Optional[str] = None,
                bbox: Optional[Tuple[float, float, float, float]] = None,
-               ecv: Optional[str] = None,
-               frequency: Optional[str] = None,
-               institute: Optional[str] = None,
-               processing_level: Optional[str] = None,
-               product_string: Optional[str] = None,
-               product_version: Optional[str] = None,
-               data_type: Optional[str] = None,
-               sensor: Optional[str] = None,
-               platform: Optional[str] = None) -> List[str]:
+               cci_attrs: Optional[Mapping[str, str]] = None) -> List[str]:
         candidate_names = []
-        if not self._data_sources and not ecv and not frequency and not processing_level \
-                and not data_type and not product_string and not product_version:
+        if not self._data_sources and 'ecv' not in cci_attrs \
+                and 'frequency' not in cci_attrs \
+                and 'processing_level' not in cci_attrs \
+                and 'data_type' not in cci_attrs \
+                and 'product_string' not in cci_attrs \
+                and 'product_version' not in cci_attrs:
             self._run_with_session(self._read_all_data_sources)
             candidate_names = self.dataset_names
         else:
             for dataset_name in self.dataset_names:
-                split_dataset_name = dataset_name.split('.')
-                if ecv is not None and ecv != split_dataset_name[1]:
+                _, ecv, frequency, processing_level, data_type, sensor, \
+                platform, product_string, product_version, _ = \
+                    dataset_name.split('.')
+                if cci_attrs.get('ecv', ecv) != ecv:
                     continue
-                if frequency is not None and frequency != \
-                        _convert_time_from_drs_id(split_dataset_name[2]):
+                if cci_attrs.get('processing_level', processing_level) \
+                        != processing_level:
                     continue
-                if processing_level is not None and processing_level != split_dataset_name[3]:
+                if cci_attrs.get('data_type', data_type) != data_type:
                     continue
-                if data_type is not None and data_type != split_dataset_name[4]:
+                if cci_attrs.get('product_string', product_string) != \
+                        product_string:
                     continue
-                if product_string is not None and product_string != split_dataset_name[7]:
+                product_version = product_version.replace('-', '.')
+                if cci_attrs.get('product_version', product_version) \
+                        != product_version:
                     continue
-                if product_version is not None and product_version.replace('.', '-') != \
-                        split_dataset_name[8]:
+                converted_time = _convert_time_from_drs_id(frequency)
+                if cci_attrs.get('frequency', converted_time) != converted_time:
                     continue
                 candidate_names.append(dataset_name)
             if len(candidate_names) == 0:
                 return []
-        if not start_date and not end_date and not institute and not sensor and not platform \
-                and not bbox:
+        if not start_date and not end_date and not bbox \
+                and 'institute' not in cci_attrs \
+                and 'sensor' not in cci_attrs \
+                and 'platform' not in cci_attrs:
             return candidate_names
         results = []
         if start_date:
@@ -651,12 +654,16 @@ class CciOdp:
             data_source_info = self._data_sources.get(candidate_name, None)
             if not data_source_info:
                 continue
-            if institute is not None and ('institute' not in data_source_info or
-                                          institute != data_source_info['institute']):
+            institute = cci_attrs.get('institute')
+            if institute is not None and \
+                    ('institute' not in data_source_info or
+                     institute != data_source_info['institute']):
                 continue
-            if sensor is not None and sensor != data_source_info['sensor_id']:
+            if cci_attrs.get('sensor', data_source_info['sensor_id']) \
+                    != data_source_info['sensor_id']:
                 continue
-            if platform is not None and platform != data_source_info['platform_id']:
+            if cci_attrs.get('platform', data_source_info['platform_id']) \
+                    != data_source_info['platform_id']:
                 continue
             if bbox:
                 if float(data_source_info['bbox_minx']) > bbox[2]: